@@ -2,7 +2,7 @@ package com.example.demo.api.controller.document;
 
 import com.example.demo.ApiDocumentationTest;
 import com.example.demo.service.dto.UnitDto;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.ResultActions;
